refactor(pdf-service): move /app/domains/share/services/pdf/attachment to /app/services/pdf/attachment
