+++ app/domains/beratungshilfe/services/pdf/pdfForm/A_angelegenheit.ts
@@ -2,7 +2,7 @@ import type { BeratungshilfeFormularContext } from "~/domains/beratungshilfe/for
 import {
   SEE_IN_ATTACHMENT_DESCRIPTION,
   type AttachmentEntries,
-} from "~/domains/shared/services/pdf/attachment";
+} from "~/services/pdf/attachment";
 import type { BerHPdfFillFunction } from "..";
 
 export const THEMA_RECHTSPROBLEM_TITLE = "Thema des Rechtsproblems";
