+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/A_person.ts
@@ -1,9 +1,9 @@
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
+import { maritalDescriptionMapping } from "~/domains/shared/services/pdf/maritalDescriptionMapping";
 import {
   type AttachmentEntries,
   SEE_IN_ATTACHMENT_DESCRIPTION,
-} from "~/domains/shared/services/pdf/attachment";
-import { maritalDescriptionMapping } from "~/domains/shared/services/pdf/maritalDescriptionMapping";
+} from "~/services/pdf/attachment";
 import type { PkhPdfFillFunction } from "..";
 
 export const GESETZLICHERVERTRETER_FIELD_MAX_CHARS = 80;
