+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/D_angehoerige.ts
@@ -1,5 +1,5 @@
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
 import { getFillUnterhalt } from "~/domains/shared/services/pdf/unterhaltHelpers";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import type { PkhPdfFillFunction } from "..";
 export const ATTACHMENT_DESCRIPTION_SECTION_D =
   "FELD D: Angehörige, denen Sie Bar- oder Naturalunterhalt gewähren";
