+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/E_bruttoEinnahmen/bruttoEinnahmen_eigenes.ts
@@ -2,7 +2,7 @@ import { finanzielleAngabeEinkuenfteGuards as guards } from "~/domains/prozessko
 import {
   type AttachmentEntries,
   SEE_IN_ATTACHMENT_DESCRIPTION,
-} from "~/domains/shared/services/pdf/attachment";
+} from "~/services/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import type { PkhPdfFillFunction } from "../..";
 import { nettoString, removeDecimalsFromCurrencyString } from "../../util";
