+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/__test__/I_zahlungsverpflichtungen.test.ts
@@ -1,7 +1,7 @@
 import type { ProzesskostenhilfePDF } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import { getProzesskostenhilfeParameters } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import { fillZahlungsverpflichtungen } from "../I_zahlungsverpflichtungen";
 
 let pdfParams: ProzesskostenhilfePDF;
