+++ app/services/pdf/attachment/createAttachmentPages.ts
@@ -1,8 +1,8 @@
 import type PDFDocument from "pdfkit";
 import type { BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import type { AttachmentEntries } from "~/domains/shared/services/pdf/attachment";
-import { createAttachmentEntries } from "~/domains/shared/services/pdf/attachment/createAttachmentEntries";
+import type { AttachmentEntries } from "~/services/pdf/attachment";
+import { createAttachmentEntries } from "~/services/pdf/attachment/createAttachmentEntries";
 import { createHeading } from "~/services/pdf/createHeading";
 import { createHeader } from "~/services/pdf/header/createHeader";
